Rename settings to environment
Make environment composable by splitting it up per registration
AB#49052
